# services/auth-go/openapi.yaml
openapi: 3.1.0
info:
  title: LeetRank Auth API (Go)
  version: 1.0.0
  summary: Authentication, session, and JWKS endpoints for LeetRank.
  description: |
    Production Go implementation of the LeetRank auth surface — owns
    register, login, me, logout, and change-password. Issues HS256-signed
    session cookies (Ed25519 in Phase 3.1.5+).

    Sliding-window rate limit: 5 attempts per 15 minutes per IP **and** per
    account on `/login` and `/register`. Bcrypt cost 12. Dummy-hash
    timing-attack mitigation on login miss.

    See [ADR 0017](../../docs/adr/0017-auth-go-rewrite.md) for the rewrite
    rationale and [ADR 0013](../../docs/adr/0013-service-to-service-auth.md)
    for the Ed25519 JWKS plan.
  contact:
    name: LeetRank Engineering
    email: jasonbmt06@gmail.com
    url: https://github.com/JasonTM17/Leetrank_Project
  license:
    name: MIT
    url: https://github.com/JasonTM17/Leetrank_Project/blob/main/LICENSE
servers:
  - url: http://localhost:4011
    description: Local dev (direct)
  - url: http://localhost:3000/api/v1/auth
    description: Local dev (via Caddy, mounted at /api/v1/auth)
  - url: https://leetrank.example.com/api/v1/auth
    description: Production
...
paths:
...
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Create an account
      description: |
        Validates email, username, and password. Hashes the password with
        bcrypt cost 12. On success, sets a `leetrank_session` cookie and
        returns the public profile. Rate-limited 5/15min per IP and per
        email.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
            examples:
              default:
                value:
                  email: alice@example.com
                  username: alice
                  password: "Sup3rSecret!"
      responses:
        "201":
          description: Account created. Session cookie set.
          headers:
            Set-Cookie:
              schema: { type: string }
              description: |
                `leetrank_session=<jwt>; Path=/; HttpOnly; SameSite=Lax;
                Secure; Max-Age=604800`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
              examples:
                default:
                  value:
                    id: "01H8X8VH8M3Q1SAMPLE0000001"
                    email: alice@example.com
                    username: alice
                    role: user
                    rating: 1500
                    createdAt: "2026-05-19T07:49:00Z"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Sign in
      description: |
        Verifies email + password (bcrypt cost 12). On hit, sets a session
        cookie. On miss, runs a dummy bcrypt comparison so timing does not
        leak whether the email exists. Rate-limited 5/15min per IP and per
        email.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
            examples:
              default:
                value:
                  email: alice@example.com
                  password: "Sup3rSecret!"
      responses:
        "200":
          description: Authenticated. Session cookie set.
          headers:
            Set-Cookie:
              schema: { type: string }
              description: |
                `leetrank_session=<jwt>; Path=/; HttpOnly; SameSite=Lax;
                Secure; Max-Age=604800`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
              examples:
                default:
                  value:
                    id: "01H8X8VH8M3Q1SAMPLE0000001"
                    email: alice@example.com
                    username: alice
                    role: user
                    rating: 1500
                    createdAt: "2026-05-19T07:49:00Z"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /v1/auth/me:
    get:
      tags: [Auth]
      summary: Inspect the current session
      description: |
        Verifies the `leetrank_session` cookie and returns the public
        profile. Returns 401 if the cookie is absent, malformed, expired,
        or signed by an unknown key.
      operationId: getCurrentUser
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current authenticated user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
              examples:
                default:
                  value:
                    id: "01H8X8VH8M3Q1SAMPLE0000001"
                    email: alice@example.com
                    username: alice
                    role: user
                    rating: 1500
                    createdAt: "2026-05-19T07:49:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      description: |
        Requires the caller to be authenticated and to provide the
        current password. Bumps the JWT key version so any other open
        sessions are invalidated.
      operationId: changePassword
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
            examples:
              default:
                value:
                  currentPassword: "Sup3rSecret!"
                  newPassword: "NewEv3nMor3Secret!"
      responses:
        "200":
          description: Password changed; new session cookie set.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, example: ok }
              examples:
                default:
                  value: { status: ok }
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/audit-log:
    get:
      tags: [Auth]
      summary: Inspect the caller's audit log
      description: |
        Returns the authenticated user's last 50 audit entries
        (login, logout, password_change, account_locked,
        profile_update, register, password_reset_requested,
        email_verified). Use ?limit=N (1..200) to override the
        default page size.
      operationId: getAuditLog
      security:
        - sessionCookie: []
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Audit log entries (most recent first).
          content:
            application/json:
              schema:
                type: object
                required: [entries, limit]
                properties:
                  limit: { type: integer, example: 50 }
                  entries:
                    type: array
                    items:
                      type: object
                      required: [id, action, createdAt]
                      properties:
                        id: { type: integer, format: int64, example: 17 }
                        action: { type: string, example: login }
                        ipAddress: { type: string, example: "203.0.113.7" }
                        userAgent: { type: string, example: "Mozilla/5.0 ..." }
                        metadata: { type: object, additionalProperties: true }
                        createdAt: { type: string, format: date-time }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password-reset link
      description: |
        Generates a single-use, time-limited (1h) password reset token
        for the supplied email and stores its sha256 hash. Always
        returns 202 to avoid leaking whether the email is registered.
        In dev mode (LEETRANK_DEV_RETURN_TOKENS=1) the plaintext token
        is echoed back in the response body for tests.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email, example: alice@example.com }
      responses:
        "202":
          description: Request accepted (whether or not the email exists).
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [accepted] }
                  resetToken:
                    type: string
                    description: Plaintext token, dev-only.
  /v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Consume a password-reset token
      description: |
        Exchanges a valid, unused, unexpired reset token for a fresh
        bcrypt hash on the matching User row. The token is marked used
        atomically with the password update.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string, example: "1q2w3e4r..." }
                newPassword: { type: string, minLength: 8, example: "NewEv3nMor3Secret!" }
      responses:
        "200":
          description: Password updated.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean, example: true }
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: Token invalid, expired, or already consumed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify the caller's email address
      description: |
        Consumes an email-verification token (24h TTL, single-use) and
        flips User.emailVerified to true.
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, example: "1q2w3e4r..." }
      responses:
        "200":
          description: Email marked verified.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean, example: true }
        "401":
          description: Token invalid, expired, or already consumed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend the email-verification link
      description: |
        Requires authentication. Issues a fresh verification token for
        the caller's account. Always returns 202.
      operationId: resendVerification
      security:
        - sessionCookie: []
      responses:
        "202":
          description: Request accepted.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [accepted] }
                  verifyToken:
                    type: string
                    description: Plaintext token, dev-only.
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: leetrank_session
      description: |
        HS256-signed JWT (Ed25519 in Phase 3.1.5+). Set by `/login` and
        `/register`, cleared by `/logout`. `HttpOnly`, `Secure`,
        `SameSite=Lax`, 7-day max-age.
  responses:
    Unauthorized:
      description: Missing, malformed, or invalid session cookie.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            default:
              value: { error: unauthorized, message: "Authentication required." }
    InvalidCredentials:
      description: Wrong email/password (rate-limited).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            default:
              value: { error: invalid_credentials, message: "Email or password is incorrect." }
    ValidationError:
      description: Request body failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
          examples:
            default:
              value:
                error: validation_failed
                message: "Validation failed."
                fields:
                  - path: password
                    message: "Password must be at least 8 characters."
    Conflict:
      description: Email or username already taken.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            email:
              value: { error: email_taken, message: "Email already in use." }
            username:
              value: { error: username_taken, message: "Username already in use." }
    RateLimited:
      description: Too many attempts in the current window.
      headers:
        Retry-After:
          schema: { type: integer, example: 537 }
          description: Seconds until the next request will be accepted.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            default:
              value:
                error: rate_limited
                message: "Too many attempts. Try again in 9 minutes."
  schemas:
...
    UserProfile:
      type: object
      required: [id, email, username, role, createdAt]
      properties:
        id: { type: string, example: "01H8X8VH8M3Q1SAMPLE0000001" }
        email: { type: string, format: email, example: alice@example.com }
        username: { type: string, example: alice }
        role: { type: string, enum: [user, admin], example: user }
        rating: { type: integer, nullable: true, example: 1500 }
        createdAt: { type: string, format: date-time, example: "2026-05-19T07:49:00Z" }
    RegisterRequest:
      type: object
      required: [email, username, password]
      properties:
        email: { type: string, format: email, example: alice@example.com }
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: "^[a-z0-9_]{3,30}$"
          example: alice
        password:
          type: string
          minLength: 8
          example: "Sup3rSecret!"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: alice@example.com }
        password: { type: string, example: "Sup3rSecret!" }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, example: "Sup3rSecret!" }
        newPassword:
          type: string
          minLength: 8
          example: "NewEv3nMor3Secret!"